# Remove the response mutex, refactor fetch_node and fix response merging