--- a/apollo-router-core/src/json_ext.rs
+++ b/apollo-router-core/src/json_ext.rs
@@ -35,6 +35,10 @@ pub trait ValueExt {
     /// values in `self`.
     #[track_caller]
     fn is_subset(&self, superset: &Value) -> bool;
+
+    /// insert a specific value at a subpath
+    #[track_caller]
+    fn insert_data(&mut self, path: &Path, value: Value) -> Result<(), FetchError>;
 }
 
 impl ValueExt for Value {
@@ -211,6 +215,27 @@ impl ValueExt for Value {
             }
             (a, b) => a == b,
         }
+    }
+
+    fn insert_data(&mut self, path: &Path, value: Value) -> Result<(), FetchError> {
+        let mut nodes =
+            self.get_at_path_mut(path)
+                .map_err(|err| FetchError::ExecutionPathNotFound {
+                    reason: err.to_string(),
+                })?;
+
+        let len = nodes.len();
+        //FIXME: are there cases where we could write at multiple paths?
+        for (i, node) in nodes.iter_mut().enumerate() {
+            if i == len {
+                (*node).deep_merge(value);
+                break;
+            } else {
+                (*node).deep_merge(value.clone());
+            }
+        }
+
+        Ok(())
     }
 }