… apollo-router-core/src/naive_introspection.rs
… apollo-router-core/src/query_planner/mod.rs
… apollo-router-core/src/query_planner/selection.rs
… apollo-router-core/src/query_planner/snapshots/apollo_router_core__query_planner__router_bridge_query_planner__tests__plan.snap
… apollo-router-core/src/query_planner/snapshots/apollo_router_core__query_planner__tests__query_plan_from_json.snap
… apollo-router-core/src/response.rs
… apollo-router/src/apollo_router.rs