Fix: #248

This PR gets rid of the response mutex and simplify response merging.

- separate tasks of `fetch_node` in different methods to extract the mutex usage more easily
- refactor `execute_recursively` to merge response data incrementally from smaller subsets of the response
- use `FuturesUnordered` to transform parallel queries in a stream of subgraph responses that are then merged one by one
- generate with representations data a path for each entity, so that we know exactly where to insert it in the response data. This fixes #248