@Geal 2021-12-07
https://app.circleci.com/pipelines/github/apollographql/router/686/workflows/d35d0548-58ed-4175-b643-4b322849d157/jobs/3854/parallel-runs/0/steps/0-113 <- that is a funny error, because the two responses are apparently the same :smile: