@Geal 2021-12-07
ok so, it passes the tests and lints, so good news. It needs a bit more cleanup now, and some perf tests, and that should be good to go