@Geal 2021-12-08
Benchmark:

## main at a296efbfde20c92ca24785b32c3b31b0636306fa

```
$ hey -m POST -H 'Content-type: application/json' -d '{"query":"query  { products(limit: 2) { review { body } re
lated { price price1 price2 price3 } } }\n","price price1 price2 price3 variables":{}}' -n 300000 -c 1000 http://127.0.0.1:4000/        
                                                                                                                                        
Summary:                                                                                                                                
  Total:        40.7823 secs                                                                                                            
  Slowest:      4.7580 secs                                                                                                             
  Fastest:      0.0009 secs                                                                                                             
  Average:      0.1238 secs                                                                                                             
  Requests/sec: 7356.1269                                                                                                               
                                                                                                                                        
                                                                                                                                        
Response time histogram:                                                                                                                
  0.001 [1]     |                                                                                                                       
  0.477 [277744]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                                                                       
  0.952 [1987]  |                                                                                                                       
  1.428 [9771]  |■                                                                                                                      
  1.904 [9931]  |■                                                                                                                      
  2.379 [37]    |                                                                                                                       
  2.855 [255]   |                                                                                                                       
  3.331 [259]   |                                                                                                                       
  3.807 [10]    |                                                                                                                       
  4.282 [3]     |                                                                                                                       
  4.758 [2]     |                                                                                                                       
                                                                                                                                        
                                                                                                                                        
Latency distribution:                                                                                                                   
  10% in 0.0139 secs                                                                                                                    
  25% in 0.0165 secs                                                                                                                    
  50% in 0.0206 secs                                                                                                                    
  75% in 0.0248 secs                                                                                                                    
  90% in 0.0301 secs                                                                                                                    
  95% in 1.3433 secs                                                                                                                    
  99% in 1.5689 secs                                                                                                                    
                                                                                                                                        
Details (average, fastest, slowest):                                                                                                    
  DNS+dialup:   0.0001 secs, 0.0009 secs, 4.7580 secs                                                                                   
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs                                                                                   
  req write:    0.0000 secs, 0.0000 secs, 0.0144 secs                                                                                   
  resp wait:    0.0270 secs, 0.0000 secs, 1.8670 secs                                                                                   
  resp read:    0.0966 secs, 0.0000 secs, 4.7502 secs                                                                                   
                                                                                                                                        
Status code distribution:                                                                                                               
  [200] 300000 responses
```

## This PR at 4229fa8fd1889499b76d4bf858f1456338c12545

```
$ hey -m POST -H 'Content-type: application/json' -d '{"query":"query  { products(limit: 2) { review { body } re
lated { price price1 price2 price3 } } }\n","price price1 price2 price3 variables":{}}' -n 300000 -c 1000 http://127.0.0.1:4000/        
                                                                                                                                        
Summary:                                                                                                                                
  Total:        38.6082 secs                                                                                                            
  Slowest:      4.4451 secs                                                                                                             
  Fastest:      0.0014 secs                                                                                                             
  Average:      0.1174 secs                                                                                                             
  Requests/sec: 7770.3751                                                                                                               
                                                                                                                                        
                                                                                                                                        
Response time histogram:                                                                                                                
  0.001 [1]     |                                                                                                                       
  0.446 [277055]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■                                                                       
  0.890 [2098]  |                                                                                                                       
  1.334 [10673] |■■                                                                                                                     
  1.779 [9558]  |■                                                                                                                      
  2.223 [48]    |                                                                                                                       
  2.668 [312]   |                                                                                                                       
  3.112 [242]   |                                                                                                                       
  3.556 [1]     |                                                                                                                       
  4.001 [8]     |                                                                                                                       
  4.445 [4]     |                                                                                                                       
                                                                                                                                        
                                                                                                                                        
Latency distribution:                                                                                                                   
  10% in 0.0132 secs                                                                                                                    
  25% in 0.0158 secs                                                                                                                    
  50% in 0.0195 secs                                                                                                                    
  75% in 0.0235 secs                                                                                                                    
  90% in 0.0280 secs                                                                                                                    
  95% in 1.2608 secs                                                                                                                    
  99% in 1.4574 secs                                                                                                                    
                                                                                                                                        
Details (average, fastest, slowest):                                                                                                    
  DNS+dialup:   0.0001 secs, 0.0014 secs, 4.4451 secs                                                                                   
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs                                                                                   
  req write:    0.0000 secs, 0.0000 secs, 0.0145 secs                                                                                   
  resp wait:    0.0251 secs, 0.0000 secs, 1.5970 secs                                                                                   
  resp read:    0.0922 secs, 0.0000 secs, 4.0642 secs                                                                                   
                                                                                                                                        
Status code distribution:                                                                                                               
  [200] 300000 responses
```

So it is slightly faster. Honestly I was expecting it to be a bit slower, because I have not started to optimize it yet, but that means it should be good to merge as is (at least on the performance side)