@Geal 2021-12-10
I ran some quick benchmarks locally, with the same query:

```graphql
query { products(limit: 2, size: 32) { price pri
ce1 price2 price3 price4 price5 price6 price7 price8 price9 price10 price11 price12 price13 price14 price15 price16 price17 price18 price19 price20 price21 price22 price23 price24 price25 price26 price27 price28 price29 price30 price31 price32 price33 price34 price35 price36 price37 price38 price39 price40 price41 price42 price43 price44 price45 price46 price47 price48 price49 price50 price51 price52 price53 price54 price55 price56 price57 price58 price59 price60 price61 price62 price63 price64 price65 price66 price67 price68 price69 price70 price71 price72 price73 price74 price75 price76 price77 price78 price79 price80 price81 price82 price83 price84 price85 price86 price87 price88 price89 price90 price91 price92 price93 price94 price95 price96 price97 price98 price99 review { product { related { review 
{ product { related { price price1 price2 price3 price4 price5 price6 price7 price8 price9 price10 price11 price12 price13 price14 price15 price16 price17 price18 price19 price20 price21 price22 price23 price24 price25 price26 price27 price28 price29 price30 price31 price32 price33 price34 price35 price36 price37 price38 price39 price40 price41 price42 price43 price44 price45 price46 price47 price48 price49 price50 price51 price52 price53 price54 price55 price56 price57 price58 price59 price60 price61 price62 price63 price64 price65 price66 price67 price68 price69 price70 price71 price72 price73 price74 price75 price76 price77 price78 price79 price80 price81 price82 price83 price84 price85 price86 price87 price88 price89 price90 price91 price92 price93 price94 price95 price96 price97 price98 price99  } } 
} } } } } }
```
with hey and 200000 queries, 1000 concurrent clients, 1 core for the router (it saturates easily like this)

On main at 4617741:

```
Summary:                                                                                                                                
  Total:        120.8086 secs                                                                                                           
  Slowest:      10.8656 secs                                                                                                            
  Fastest:      0.0056 secs                                                                                                             
  Average:      0.5549 secs                                                                                                             
  Requests/sec: 1655.5115                                                                                                                                                                                                                           
                                                                                                                                   
Latency distribution:                                                                                                                   
  10% in 0.0445 secs                                                                                                                    
  25% in 0.0553 secs                                                                                                                    
  50% in 0.0739 secs                                                                                                                    
  75% in 0.0940 secs                                                                                                                    
  90% in 2.8722 secs                                                                                                                    
  95% in 3.0867 secs                                                                                                                    
  99% in 5.2827 secs    
```

This PR at bf9e5bb:

```
Summary:                                                          
  Total:        116.1452 secs                                     
  Slowest:      10.6822 secs                                      
  Fastest:      0.0053 secs                                       
  Average:      0.5313 secs                                       
  Requests/sec: 1721.9827                                         
                                                          
Latency distribution:                                             
  10% in 0.0427 secs                                              
  25% in 0.0533 secs                                              
  50% in 0.0711 secs                                              
  75% in 0.0908 secs                                              
  90% in 2.7576 secs                                              
  95% in 2.9643 secs                                              
  99% in 5.1863 secs
```

This PR at db4a518:

```
Summary:                                                              
  Total:        120.5513 secs                                         
  Slowest:      13.7664 secs                                          
  Fastest:      0.0062 secs                                           
  Average:      0.5552 secs                                           
  Requests/sec: 1659.0443                                             
                                                                
Latency distribution:                                                 
  10% in 0.0438 secs                                                  
  25% in 0.0550 secs                                                  
  50% in 0.0738 secs                                                  
  75% in 0.0941 secs                                                  
  90% in 2.9021 secs                                                  
  95% in 3.1246 secs                                                  
  99% in 5.1338 secs
```

So the results are inconclusive for now, but I see potential for improvement in the profiling traces